IA-16: fix some type mismatches so that it compiles with DJGPP
(cherry picked from commit 70ff0a869edcabdf00cde547dc5f0ab9b50325c1)
(cherry picked from commit d04c6c2efdc28b13383fbbecd1f9e1ece6ec0e7f)

+++ binutils/objdump.c
@@ -1890,7 +1890,7 @@ update_source_path (const char *filename, bfd *abfd)
     {
       long mtime = bfd_get_mtime (abfd);
 
-      if (fst.st_mtime > mtime)
+      if ((long) fst.st_mtime > mtime)
 	warn (_("source file %s is more recent than object file\n"),
 	      filename);
     }
